Running maximum

min_run calculates running max on given x numeric vector, specified k window size.

max_run(
  x,
  k = integer(0),
  lag = integer(1),
  idx = integer(0),
  at = integer(0),
  na_rm = TRUE,
  na_pad = FALSE
)

Arguments

x
(vector, data.frame, matrix, xts)
Input in runner custom function f.

k
(integer vector or single value)
Denoting size of the running window. If k is a single value then window size is constant for all elements, otherwise if length(k) == length(x) different window size for each element. One can also specify k in the same way as by argument in seq.POSIXt. See 'Specifying time-intervals' in details section.

lag
(integer vector or single value)
Denoting window lag. If lag is a single value then window lag is constant for all elements, otherwise if length(lag) == length(x) different window size for each element. Negative value shifts window forward. One can also specify lag in the same way as by argument in seq.POSIXt. See 'Specifying time-intervals' in details section.

idx
(integer, Date, POSIXt)
Optional integer vector containing sorted (ascending) index of observation. By default idx is index incremented by one. User can provide index with varying increment and with duplicated values. If specified then k and lag are depending on idx. Length of idx have to be equal of length x.

at
(integer, Date, POSIXt, character vector)
Vector of any size and any value defining output data points. Values of the vector defines the indexes which data is computed at. Can be also POSIXt sequence increment used in at argument in seq.POSIXt. See 'Specifying time-intervals' in details section.

na_rm
logical single value (default na_rm = TRUE) - if TRUE sum is calculating excluding NA.

na_pad
(logical single value)
Whether incomplete window should return NA (if na_pad = TRUE) Incomplete window is when some parts of the window are out of range.

Value

max numeric vector of length equals length of x.

Examples

set.seed(11)
x1 <- sample( c(1,2,3), 15, replace=TRUE)
x2 <- sample( c(NA,1,2,3), 15, replace=TRUE)
k  <- sample( 1:4, 15, replace=TRUE)
max_run(x1) # simple cumulative maximum

#>  [1] 2 2 2 2 2 2 2 2 2 2 3 3 3 3 3

max_run(x2, na_rm = TRUE) # cumulative maximum with removing NA.

#>  [1] 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3

max_run(x2, na_rm = TRUE, k=4) # maximum in 4-element window

#>  [1] 2 2 2 2 2 2 3 3 3 3 2 2 2 2 2

max_run(x2, na_rm = FALSE, k=k) # maximum in varying k window size

#>  [1]   2   2 NaN NaN NaN NaN   3   3   2   1   2   2   2   2   2