Set window parameters

Set window parameters for runner. This function sets the attributes to x (only data.frame) object and saves user effort to specify window parameters in further multiple runner calls.

run_by(x, idx, k, lag, na_pad, at)

Arguments

x

(vector, data.frame, matrix, xts, grouped_df)
Input in runner custom function f.

idx

(integer, Date, POSIXt)
Optional integer vector containing sorted (ascending) index of observation. By default idx is index incremented by one. User can provide index with varying increment and with duplicated values. If specified then k and lag are depending on idx. Length of idx have to be equal of length x.

k

(integer vector or single value)
Denoting size of the running window. If k is a single value then window size is constant for all elements, otherwise if length(k) == length(x) different window size for each element. One can also specify k in the same way as by argument in seq.POSIXt. See 'Specifying time-intervals' in details section.

lag

(integer vector or single value)
Denoting window lag. If lag is a single value then window lag is constant for all elements, otherwise if length(lag) == length(x) different window size for each element. Negative value shifts window forward. One can also specify lag in the same way as by argument in seq.POSIXt. See 'Specifying time-intervals' in details section.

na_pad

(logical single value)
Whether incomplete window should return NA (if na_pad = TRUE) Incomplete window is when some parts of the window are out of range.

at

(integer, Date, POSIXt, character vector)
Vector of any size and any value defining output data points. Values of the vector defines the indexes which data is computed at. Can be also POSIXt sequence increment used in at argument in seq.POSIXt. See 'Specifying time-intervals' in details section.

Value

x object which runner can be executed on.

Examples

if (FALSE) {
library(dplyr)

data <- data.frame(
 index = c(2, 3, 3, 4, 5, 8, 10, 10, 13, 15),
 a = rep(c("a", "b"), each = 5),
 b = 1:10
)

data %>%
 group_by(a) %>%
 run_by(idx = "index", k = 5) %>%
 mutate(
   c = runner(
     x = .,
     f = function(x) {
       paste(x$b, collapse = ">")
     }
   ),
   d = runner(
     x = .,
     f = function(x) {
       sum(x$b)
     }
   )
 )
}