Custom running function

Applies custom function to running windows

runner(x, f, k = integer(1), lag = integer(1), idx = integer(0))

Arguments

x
Vector of any type

f
function to be applied on x

k
integer vector or single value denoting size of the running window. If k is a single value then window size is constant for all elements, otherwise if length(k) == length(x) different window size for each element.

lag
integer vector or single value denoting window lag. If lag is a single value then window lag is constant for all elements, otherwise if length(lag) == length(x) different window size for each element.

idx
date or integer an optional integer vector containing index of observation. If specified then k and lag are depending on idx. Length of idx should be equal of length x

Examples

runner(1:10, f = mean, k = 3)

#>  [1] 1.0 1.5 2.0 3.0 4.0 5.0 6.0 7.0 8.0 9.0

runner(1:10, k = 3, f = function(x) mean(x, na.rm = TRUE))

#>  [1] 1.0 1.5 2.0 3.0 4.0 5.0 6.0 7.0 8.0 9.0

runner(letters[1:10],
       k = c(1, 2, 2, 4, 5, 5, 5, 5, 5, 5),
       f = function(x) length(unique(x)))

#>  [1] 1 2 2 4 5 5 5 5 5 5