List of running windows

Creates list of windows

window_run(x, k = integer(1), lag = integer(1), idx = integer(0),
  na_pad = FALSE)

Arguments

x
Vector of any type

k
integer vector or single value denoting size of the running window. If k is a single value then window size is constant for all elements, otherwise if length(k) == length(x) different window size for each element.

lag
integer vector or single value denoting window lag. If lag is a single value then window lag is constant for all elements, otherwise if length(lag) == length(x) different window size for each element.

idx
date or integer an optional integer vector containing index of observation. If specified then k and lag are depending on idx. Length of idx should be equal of length x

na_pad
logical single value (default na_pad=FALSE) - if TRUE calculation on incomplete window will return NA. Incomplete window is when some parts of the window are out of range

Examples

window_run(1:10, k = 3, lag = -1)

#> [[1]]
#> [1] 1 2
#> 
#> [[2]]
#> [1] 1 2 3
#> 
#> [[3]]
#> [1] 2 3 4
#> 
#> [[4]]
#> [1] 3 4 5
#> 
#> [[5]]
#> [1] 4 5 6
#> 
#> [[6]]
#> [1] 5 6 7
#> 
#> [[7]]
#> [1] 6 7 8
#> 
#> [[8]]
#> [1] 7 8 9
#> 
#> [[9]]
#> [1]  8  9 10
#> 
#> [[10]]
#> [1]  9 10
#> 

window_run(letters[1:10], k = c(1, 2, 2, 4, 5, 5, 5, 5, 5, 5))

#> [[1]]
#> [1] "a"
#> 
#> [[2]]
#> [1] "a" "b"
#> 
#> [[3]]
#> [1] "b" "c"
#> 
#> [[4]]
#> [1] "a" "b" "c" "d"
#> 
#> [[5]]
#> [1] "a" "b" "c" "d" "e"
#> 
#> [[6]]
#> [1] "b" "c" "d" "e" "f"
#> 
#> [[7]]
#> [1] "c" "d" "e" "f" "g"
#> 
#> [[8]]
#> [1] "d" "e" "f" "g" "h"
#> 
#> [[9]]
#> [1] "e" "f" "g" "h" "i"
#> 
#> [[10]]
#> [1] "f" "g" "h" "i" "j"
#>